+++ src/sdk/PnP.Core.Test/SharePoint/WebTests.cs
@@ -2239,6 +2239,17 @@ namespace PnP.Core.Test.SharePoint
             }
         }
 
+        [TestMethod]
+        public async Task GetEffectiveExternalUserPermissionsAsyncTest()
+        {
+            TestCommon.Instance.Mocking = false;
+            using (var context = await TestCommon.Instance.GetContextAsync(TestCommon.TestSite))
+            {
+                var basePermissions = await context.Web.GetUserEffectivePermissionsAsync("bjansen_microsoft.com#ext#@bertonline.onmicrosoft.com");
+
+                Assert.IsNotNull(basePermissions);
+            }
+        }
 
         [TestMethod]
         public async Task CheckIfUserHasPermissionsAsyncTest()